WebSettings: Refactor masking screen
