+++ TF_Settings_Web/package.json
@@ -58,7 +58,6 @@
         "eslint": "^8.18.0",
         "eslint-plugin-no-relative-import-paths": "^1.3.4",
         "eslint-plugin-react": "^7.30.0",
-        "eslint-plugin-unicorn": "^42.0.0",
         "prettier": "^2.7.1",
         "prettier-eslint": "^15.0.1",
         "react-router-dom": "^6.3.0",
