+++ TF_Settings_Web/src/Components/Pages/Camera/CameraManager.tsx
@@ -1,8 +1,8 @@
 import { Route, Routes } from 'react-router-dom';
 
-import CameraMaskingScreen from './CameraMaskingScreen';
 import CameraSetupScreen from './CameraSetupScreen';
 import { ManualSetupScreen } from './ManualSetupScreen';
+import MaskingScreen from './Masking/MaskingScreen';
 import QuickSetupManager from './QuickSetup/QuickSetupManager';
 
 const CameraManager = () => {
@@ -11,7 +11,7 @@ const CameraManager = () => {
             <Route path="" element={<CameraSetupScreen />} />
             <Route path="quick/*" element={<QuickSetupManager />} />
             <Route path="manual" element={<ManualSetupScreen />} />
-            <Route path="masking" element={<CameraMaskingScreen />} />
+            <Route path="masking" element={<MaskingScreen />} />
         </Routes>
     );
 };
